requirements.txt
streamlit
numpy
torch
torchvision
matplotlib
pillow
h5py
